1.2-LSTM_sentiment_classification/pyproject.toml
[project]
name = "sentiment-classification"
version = "0.1.0"
description = "LSTM Sentiment Classification on IMDB Reviews"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "gensim>=4.3.3",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "tensorflow>=2.20.0",
]

[project.scripts]
sentiment-cli = "sentiment_classification.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sentiment_classification"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "ruff>=0.14.6",
]
